+++ src/main/java/com/dyl/o2o/common/ResultCode.java
@@ -18,7 +18,6 @@ public enum ResultCode {
     //内部错误
     NO_RESULT(50000,"无相关数据"),
     INNER_ERROR(50001,"内部错误"),
-
     NO_AUTHORITY(50002,"无操作权限"),
     IO_ERROR(50003,"数据传输异常"),
     NO_LOG_IN(50004,"未登陆"),
