+++ src/main/resources/application-dev.yml
@@ -47,6 +47,6 @@ jwt:
 #设置日志相关
 logging:
   level:
-    root: info
+    root: debug
     com.dyl.o2o.dao: debug
   path: ./log
